Added Listeners and Shared preference
Added button listeners and Shared preferences for the fragments so that the app knows what fragments/activities to open for the user in future use, also added a ChildSignUpFragment and changed the current SignUpFragement to ParentSignUpFragment. Depending on the users selections in the into they will be shown a Parent Sign Up Sheet or a Child Sign Up Sheet.