Added location service
Added background service to run during the life of the app. It gets the either the parents or childs location and pushes the information to the JSON library holding the values. Also added a broadcastReciever to check if the child is within the parent set radius and will change the parent activity background and image view depending on whether or not the child is in the radius.