Submitting Project to Instructor
This is the end of this project, for now. This project is not even close to being finished, it was meant as a platform to learn several techniques in android. I may come back to this later to add legitimate user creation through firebase or AWS, rather than just plain text JSON user credentials, which is unquestionably a massive security risk. Also, The the user needs to be restricted from going to the next page in the intro until an option has been selected and the app should automatically go to the next page when an option has been selected. There also needs to be much better error handling regarding location for the user, regardless of child or parent. The app does not check if the user has location turned on and it does not handle the case where the user selects Deny when being prompted for permissions. Would also like to add estimated address viewable to the user instead of Latitude and Longitude. Likewise, I would like the user to be able to enter an address instead of latitude and longitude for specifying a location.’